Move BlocksRegister to corresponding package
